Extract react flow tree in a component
