+++ src/components/network-modification-tree.js
@@ -5,27 +5,13 @@
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
 import React, { useCallback, useEffect, useRef, useState } from 'react';
-import ReactFlow, {
-    Controls,
-    useStoreState,
-    useZoomPanHelper,
-} from 'react-flow-renderer';
-import { useDispatch, useSelector } from 'react-redux';
-import {
-    workingTreeNode,
-    selectTreeNode,
-    setModificationsDrawerOpen,
-} from '../redux/actions';
-import { useNodeSingleAndDoubleClick } from './graph/util/node-single-double-click-hook';
-import NetworkModificationNode from './graph/nodes/network-modification-node';
-import ModelNode from './graph/nodes/model-node';
-import RootNode from './graph/nodes/root-node';
+import { useStoreState } from 'react-flow-renderer';
+import { useSelector } from 'react-redux';
 import CreateNodeMenu from './graph/menus/create-node-menu';
 import { Box } from '@mui/material';
-import { createTreeNode, deleteTreeNode, buildNode } from '../utils/rest-api';
+import { createTreeNode, deleteTreeNode } from '../utils/rest-api';
 import { displayErrorMessageWithSnackbar, useIntlRef } from '../utils/messages';
 import { useSnackbar } from 'notistack';
-import CenterGraphButton from './graph/util/center-graph-button';
 import { useParams } from 'react-router-dom';
 import NodeEditor from './graph/menus/node-editor';
 import { StudyDrawer } from './study-drawer';
@@ -33,17 +19,7 @@ import makeStyles from '@mui/styles/makeStyles';
 import { DRAWER_NODE_EDITOR_WIDTH } from './map-lateral-drawers';
 import PropTypes from 'prop-types';
 import { StudyDisplayMode } from './study-pane';
-
-const nodeTypes = {
-    ROOT: RootNode,
-    NETWORK_MODIFICATION: NetworkModificationNode,
-    MODEL: ModelNode,
-};
-
-// snapGrid value set to [15, 15] which is the default value for ReactFlow
-// it has to be explicitly set as prop of the ReactFlow component, even if snapToGrid option is set to false
-// in order to avoid unwanted tree nodes rendering (react-flow bug ?)
-const snapGrid = [15, 15];
+import ReactFlowTree from './react-flow-tree';
 
 const useStyles = makeStyles((theme) => ({
     nodeEditor: {
@@ -66,14 +42,6 @@ const useStyles = makeStyles((theme) => ({
         marginLeft: -DRAWER_NODE_EDITOR_WIDTH,
     },
     container: { width: '100%', height: '100%' },
-    controls: {
-        position: 'absolute',
-        top: '10px',
-        right: '10px',
-        // Unset default properties of ReactFlow controls
-        left: 'unset',
-        bottom: 'unset',
-    },
 }));
 
 // We need the previous display and width to compute the transformation we will apply to the tree in order to keep the same focus.
@@ -88,12 +56,8 @@ const usePreviousTreeDisplay = (display, width) => {
     return ref.current;
 };
 
-const NetworkModificationTree = ({ treeModel, studyMapTreeDisplay }) => {
-    const dispatch = useDispatch();
-
+const NetworkModificationTree = ({ studyMapTreeDisplay }) => {
     const selectedNode = useSelector((state) => state.selectedTreeNode);
-
-    const [isMoving, setIsMoving] = useState(false);
 
     const intlRef = useIntlRef();
 
@@ -106,64 +70,6 @@ const NetworkModificationTree = ({ treeModel, studyMapTreeDisplay }) => {
     const isModificationsDrawerOpen = useSelector(
         (state) => state.isModificationsDrawerOpen
     );
-
-    const onElementClick = useCallback(
-        (event, element) => {
-            dispatch(
-                setModificationsDrawerOpen(
-                    element.type === 'NETWORK_MODIFICATION'
-                )
-            );
-            dispatch(selectTreeNode(element));
-            if (
-                element.type === 'ROOT' ||
-                (element.type === 'MODEL' &&
-                    element.data.buildStatus === 'BUILT')
-            ) {
-                dispatch(workingTreeNode(element));
-            }
-        },
-        [dispatch]
-    );
-
-    const onNodeDoubleClick = useCallback(
-        (event, node) => {
-            if (
-                node.type === 'MODEL' &&
-                node.data.buildStatus !== 'BUILT' &&
-                node.data.buildStatus !== 'BUILDING'
-            ) {
-                buildNode(studyUuid, node.id).catch((errorMessage) => {
-                    displayErrorMessageWithSnackbar({
-                        errorMessage: errorMessage,
-                        enqueueSnackbar: enqueueSnackbar,
-                        headerMessage: {
-                            headerMessageId: 'NodeBuildingError',
-                            intlRef: intlRef,
-                        },
-                    });
-                });
-            }
-        },
-        [studyUuid, enqueueSnackbar, intlRef]
-    );
-
-    const nodeSingleOrDoubleClick = useNodeSingleAndDoubleClick(
-        onElementClick,
-        onNodeDoubleClick
-    );
-
-    const onPaneClick = useCallback(() => {
-        dispatch(selectTreeNode(null));
-    }, [dispatch]);
-
-    const onMove = useCallback((flowTransform) => {
-        setIsMoving(true);
-    }, []);
-
-    const onMoveEnd = useCallback((flowTransform) => {
-        setIsMoving(false);
-    }, []);
 
     const handleCreateNode = useCallback(
         (element, type, insertMode) => {
@@ -222,73 +128,9 @@ const NetworkModificationTree = ({ treeModel, studyMapTreeDisplay }) => {
         });
     }, []);
 
-    const [x, y, zoom] = useStoreState((state) => state.transform);
-
-    const { transform } = useZoomPanHelper();
-
     const width = useStoreState((state) => state.width);
 
     const prevTreeDisplay = usePreviousTreeDisplay(studyMapTreeDisplay, width);
-
-    //We want to trigger the following useEffect that manage the modification tree focus only when we change the study map/tree display.
-    //So we use this useRef to avoid to trigger on those depedencies.
-    const focusParams = useRef();
-    focusParams.current = {
-        x,
-        y,
-        zoom,
-        transform,
-        prevTreeDisplay,
-    };
-
-    useEffect(() => {
-        const nodeEditorShift = isModificationsDrawerOpen
-            ? DRAWER_NODE_EDITOR_WIDTH
-            : 0;
-        const { x, y, zoom, transform, prevTreeDisplay } = focusParams.current;
-        if (prevTreeDisplay) {
-            if (
-                prevTreeDisplay.display === StudyDisplayMode.TREE &&
-                studyMapTreeDisplay === StudyDisplayMode.HYBRID
-            ) {
-                transform({
-                    x: x - (prevTreeDisplay.width + nodeEditorShift) / 4,
-                    y: y,
-                    zoom: zoom,
-                });
-            } else if (
-                prevTreeDisplay.display === StudyDisplayMode.HYBRID &&
-                studyMapTreeDisplay === StudyDisplayMode.TREE
-            ) {
-                transform({
-                    x: x + (prevTreeDisplay.width + nodeEditorShift) / 2,
-                    y: y,
-                    zoom: zoom,
-                });
-            }
-        }
-    }, [studyMapTreeDisplay, isModificationsDrawerOpen]);
-
-    useEffect(() => {
-        const { x, y, zoom, transform } = focusParams.current;
-        if (isModificationsDrawerOpen) {
-            transform({
-                x: x - DRAWER_NODE_EDITOR_WIDTH / 2,
-                y: y,
-                zoom: zoom,
-            });
-        } else {
-            transform({
-                x: x + DRAWER_NODE_EDITOR_WIDTH / 2,
-                y: y,
-                zoom: zoom,
-            });
-        }
-    }, [isModificationsDrawerOpen]);
-
-    const onLoad = useCallback((reactFlowInstance) => {
-        reactFlowInstance.fitView();
-    }, []);
 
     return (
         <>
@@ -297,36 +139,14 @@ const NetworkModificationTree = ({ treeModel, studyMapTreeDisplay }) => {
                 display="flex"
                 flexDirection="row"
             >
-                <Box flexGrow={1}>
-                    <ReactFlow
-                        style={{ cursor: isMoving ? 'grabbing' : 'grab' }}
-                        elements={treeModel ? treeModel.treeElements : []}
-                        onNodeContextMenu={onNodeContextMenu}
-                        onElementClick={nodeSingleOrDoubleClick}
-                        onPaneClick={onPaneClick}
-                        onMove={onMove}
-                        onLoad={onLoad}
-                        onMoveEnd={onMoveEnd}
-                        elementsSelectable
-                        selectNodesOnDrag={false}
-                        nodeTypes={nodeTypes}
-                        connectionLineType="smoothstep"
-                        nodesDraggable={false}
-                        nodesConnectable={false}
-                        snapToGrid={false}
-                        // Although snapToGrid is set to false, we have to set snapGrid constant
-                        // value in order to avoid useless re-rendering
-                        snapGrid={snapGrid}
-                    >
-                        <Controls
-                            className={classes.controls}
-                            showZoom={false}
-                            showInteractive={false}
-                        >
-                            <CenterGraphButton selectedNode={selectedNode} />
-                        </Controls>
-                    </ReactFlow>
-                </Box>
+                <ReactFlowTree
+                    onNodeContextMenu={onNodeContextMenu}
+                    studyUuid={studyUuid}
+                    studyMapTreeDisplay={studyMapTreeDisplay}
+                    isModificationsDrawerOpen={isModificationsDrawerOpen}
+                    prevTreeDisplay={prevTreeDisplay}
+                />
+
                 {selectedNode && selectedNode.type === 'NETWORK_MODIFICATION' && (
                     <StudyDrawer
                         open={isModificationsDrawerOpen}
@@ -358,6 +178,5 @@ const NetworkModificationTree = ({ treeModel, studyMapTreeDisplay }) => {
 export default NetworkModificationTree;
 
 NetworkModificationTree.propTypes = {
-    treeModel: PropTypes.object,
     studyMapTreeDisplay: PropTypes.string.isRequired,
 };
